Cartesian Item Pairings List
Attached is the code I wrote in R used to take our cleaned dataset to locate all possible item pairings within that dataset, and export a full list of all of this pairings into CSV format. This file was used heavily by out team for developing our final Tableau and Power BI visualizations shown in the Presentation file in this repository.